Refactor extract errors module
