+++ lib/simple_token_authentication.rb
@@ -1,30 +1,10 @@
 require 'simple_token_authentication/acts_as_token_authenticatable'
 require 'simple_token_authentication/acts_as_token_authentication_handler'
 require 'simple_token_authentication/configuration'
+require 'simple_token_authentication/errors'
 
 module SimpleTokenAuthentication
   extend Configuration
-
-  class NoAdapterAvailableError < LoadError
-
-    def to_s
-      message = <<-HELP.gsub(/^ {8}/, '')
-        No adapter could be loaded, probably because of unavailable dependencies.
-
-        Please make sure that Simple Token Authentication is declared after your adapters' dependencies in your Gemfile.
-
-        Example:
-
-            # Gemfile
-
-            gem 'mongoid', '~> 4.0' # for example
-            gem 'simple_token_authentication', '~> 1.0'
-
-        See https://github.com/gonzalo-bulnes/simple_token_authentication/issues/158
-      HELP
-    end
-  end
-  InvalidOptionValue = Class.new(RuntimeError)
 
   private
